requirements.txt
Django==3.1.2
django-adminplus==0.5
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-grappelli==2.14.2
django-mathfilters==1.0.0
django-simple-history==2.12.0
djangorestframework
scikit-learn
sklearn
pandas